### 2019-11-07 #71 @whedon
[PRE REVIEW]: General relativity in R: visual representation of Schwarzschild space using different coordinate systems
**Submitting author:** @RobinHankin (<a href="http://orcid.org/0000-0001-5982-0415">Robin Hankin</a>)
**Repository:** <a href="https://github.com/RobinHankin/schwarzschild" target ="_blank">https://github.com/RobinHankin/schwarzschild</a>
**Version:** 1.0-1
