### 2019-11-08 #71 @labarba
👋  @RobinHankin — the paper mentions that the software is under GPL, but there's no LICENSE file in the repository. Could you add that?
