### 2019-11-08 #71 @RobinHankin
now added, sorry about that. I thought it was understood to be under GPL.

<hankin.robin@gmail.com>
hankin.robin@gmail.com

<hankin.robin@gmail.com>

<hankin.robin@gmail.com>

On Fri, Nov 8, 2019 at 1:41 PM Lorena A. Barba <notifications@github.com>
wrote:

> 👋 @RobinHankin <https://github.com/RobinHankin> — the paper mentions
> that the software is under GPL, but there's no LICENSE file in the
> repository. Could you add that?
>
