### 2019-12-13 #71 @IanHawke
I am closing in on reviewers (most GR people are claiming no knowledge of R), and hope to twist some arms in person at a conference next week.
