### 2019-12-20 #71 @RobinHankin
thanks for this Ian.  I appreciate the difficulty of finding an
intersection between R and GR (!) but this is one reason why I wrote the
package.

Best wishes for 2020

Robin
<hankin.robin@gmail.com>
hankin.robin@gmail.com

<hankin.robin@gmail.com>

<hankin.robin@gmail.com>
