### 2020-02-16 e65399d
add ellipsis to GW()
