+++ R/kruskal_with_throw.R
@@ -217,7 +217,6 @@
   ## plot the AUT logo:
   if(!isFALSE(getOption("schwarzschild_logo"))){logo(x=0.84,y=0.08, width=0.1)}  
 
-  par(family="mono")
   git(-2,-2)
 
 }
