+++ R/eddington_finkelstein_outgoing.R
@@ -107,7 +107,6 @@ eddington_outgoing <-  function(colours=standard_colours, ...){
   ## plot the AUT logo:
   if(!isFALSE(getOption("schwarzschild_logo"))){logo(x=0.8,y=0.08, width=0.1)}  
 
-  par(family="mono")
   git(-0.8,-0.7)
 
 }
