+++ R/eddington_finkelstein.R
@@ -103,7 +103,6 @@
   ## plot the AUT logo:
   if(!isFALSE(getOption("schwarzschild_logo"))){logo(x=0.8,y=0.08, width=0.1)}  
 
-  par(family="mono")
   git(-0.8,-0.7)
 
 }
