+++ R/logo.R
@@ -11,6 +11,5 @@
 }
 
 `git` <- function(x,y,...){
-    par(family="mono")
-    text(x,y,'https://github.com/RobinHankin/schwarzschild.git',pos=4,cex=0.6,...)
+    text(x,y,'https://github.com/RobinHankin/schwarzschild.git',pos=4,cex=0.6,family="mono", ...)
 }
