+++ README.Rmd
@@ -54,7 +54,7 @@ provided.
 For more detail, see the package vignette
 
 ```
-vignette("lorentz")
+vignette("schwarzschild")
 ``` 
  
 
