### 2020-09-14 #71 @atrisovic
Hi Robin - I guess this is the beauty of open review - I was checking out this repository and I think it would be great if you had some examples of diagrams (visual representations) in your paper! :-))
