### 2020-09-15 #71 @atrisovic
Hi Robin,

Here are my initial thoughts:

- I don't mean specifically Minkowski diagram, but this is a paper on visual representation, so it would be helpful to see figures and diagrams that the package can produce in a few examples
- The paper doesn't say what R version the package is compatible with. This is important because R is not normally backward compatible. For example, can I run it with R 4.0?
- The paper mentions rich documentation for the code, but there is no link to the documentation
