### 2020-09-15 #71 @RobinHankin
* have you seen the vignette? this has 19 figures in it.   Just type `vignette("schwarzschild")` at the commandline
* The version of R is not crucial, it will work with any verson of R >=3.0
* The documentation is in standard Rd format, accessible through the usual R help mechanism

HTH

Robin
