### 2020-09-22 #71 @atrisovic
Awesome! Thank you, Kyle! I think this is an interesting topic, and it would certainly warrant more text in the paper (this [previously published paper](https://jose.theoj.org/papers/10.21105/jose.00065) could be a good example).

No time to waste, so here are a few more thoughts:
- Mathematica is typically used for creating these diagrams. Are there any advantages to using this R package instead, and what are those?
- There is a note on how the package was used, but it would also be helpful to see who the intended user group is? Is it the teachers and the undergrads? Students of astrophysics and HEP? Can the tool be used for some more advanced research?
