### 2020-09-25 #91 @IanHawke
To keep all comments in one issue, the initial points that have been made are:

@atrisovic says:

[x] I don't mean specifically Minkowski diagram, but this is a paper on visual representation, so it would be helpful to see figures and diagrams that the package can produce in a few examples (*answered*: Just type `vignette("schwarzschild")` at the commandline)
[x] The paper doesn't say what R version the package is compatible with. This is important because R is not normally backward compatible. For example, can I run it with R 4.0? (*answered*: it will work with any verson of R >=3.0)
[x] The paper mentions rich documentation for the code, but there is no link to the documentation (*answered*: The documentation is in standard Rd format, accessible through the usual R help mechanism)
[ ] Mathematica is typically used for creating these diagrams. Are there any advantages to using this R package instead, and what are those?
[ ] There is a note on how the package was used, but it would also be helpful to see who the intended user group is? Is it the teachers and the undergrads? Students of astrophysics and HEP? Can the tool be used for some more advanced research?
