### 2020-10-01 #91 @atrisovic
Hi @RobinHankin,

I still stand behind my previous comments, but here is also a more official review based on the checklist above.
On the package:

- There is no tagged version of the package on GitHub.
- There are no instructions for contributing to the package. Consider adding a `contributing.md` file.
- There doesn't seem to be any tests in the package.
- I was initially unable to see the vignette, so I suggest changing the instructions command to: `devtools::install_github("RobinHankin/schwarzschild", build_vignettes = TRUE)`.

On the paper:

- The introduction does not introduce the paper's main contribution, which is an educational package for R
- The paper does not clearly state the need for this software and who the target audience is.
- The functionality and usage section could also be improved. Some of the sentences seem incomplete, like "heavy documentation" but no note that it's internal to R; "twenty functions" but no examples of any of them; "code is maintainable and extensible" but no comment on how or what would be potential further work.
- Consider adding a few diagrams in the paper. These would be helpful for the users who don't want to install the library to see the diagrams. Alternatively, if the vignette is available outside R, consider adding a link to it.
- A note on primary R version compatibility would be useful.
- Stylistic: some of the sentences might be too informal or exaggerated for an academic paper.

I am happy to have another look when the paper and the repository are updated.

Also, @IanHawke, thanks for kicking off the review. I'm not sure how to evaluate documentation and examples, as both are internal to R. It would be useful to have external examples to give the user a head start, but there are internal instructions for those who already installed the library.
