### 2020-10-05 RobinHankin/schwarzschild#14 @RobinHankin
JOSE reviews
There are a number of comments on the JOSE submission at:

https://github.com/openjournals/jose-reviews/issues/91

I am going to create one issue per bullet point with names like **JOSE review tagged version** and deal with them one by one.
