+++ contributing.md
@@ -1,0 +1,53 @@
+# Contributing to the schwarzschild package
+
+I would love your input! I want to make contributing to this project
+as easy and transparent as possible, whether it's:
+
+- Reporting a bug
+- Discussing the current state of the code
+- Submitting a fix
+- Proposing new features
+- Implementing new features
+- Reporting positive feedback and encouragement from students of GR
+- Reporting negative feedback or suggestions for new functionality
+
+I try to [assume good faith](https://en.wikipedia.org/wiki/Wikipedia:Assume_good_faith).
+
+##  Develop with Github
+
+I use github to host code, to track issues and feature requests, as
+well as accept pull requests.
+
+## I use [Github Flow](https://guides.github.com/introduction/flow/index.html), so all code changes happen through pull requests
+
+Pull requests are the best way to propose changes to the codebase (we
+use [Github
+Flow](https://guides.github.com/introduction/flow/index.html)).
+Please note that I am not 100% confident about PR etiquette, as I am
+still learning github norms.  But, I actively welcome your pull
+requests:
+
+1. Fork the repo and create your branch from `master`
+2. If you've added code that should be tested, add tests
+3. Document the added code as fully as possible
+3. Ensure the test suite passes `R CMD check`
+4. Issue that pull request!
+
+## Any contributions you make will be under the GPL Software License
+
+In short, when you submit code changes, your submissions are
+understood to be under the same that covers the project.  Feel free to
+contact the maintainers if that's a concern.
+
+## Report issues using Github's [issues](https://github.com/RobinHankin/clock/issues)
+
+I am using GitHub issues to track public bugs.  Report a bug by [opening a new issue](); it's that easy!
+
+## Use a Consistent Coding Style
+
+Doesn't really matter what style that is, so long as it's internally consistent.
+
+
+## References
+
+This document was adapted from `briandk\CONTRIBUTING.md`.  Thanks, Brian!
