+++ tests/testthat.R
@@ -1,0 +1,2 @@
+library("testthat")
+test_check("schwarzschild")
