+++ DESCRIPTION
@@ -13,7 +13,7 @@ Description:
   package.
 Depends: gsl,deSolve,animation,grImport
 License: GPL (>= 2)
-Suggests: knitr, rmarkdown
+Suggests: knitr, rmarkdown, testthat
 LazyLoad: yes
 VignetteBuilder: knitr
 URL: https://github.com/RobinHankin/schwarzschild.git
