+++ README.Rmd
@@ -32,10 +32,11 @@ images showing different aspects of the physics of black holes.
 
 # Installation
 
-The package is not yet on CRAN.  To install the current development version use `devtools`:
+The package is not yet on CRAN.  To install the current development
+version use `devtools`:
 
 ```
-R> devtools::install_github("RobinHankin/schwarzschild")
+R> devtools::install_github("RobinHankin/schwarzschild", build_vignettes=TRUE)
 ```
 
 And then to load the package use `library()`:
