### 2020-10-06 3041382
include commands to create correctly-sized jpg images
