### 2020-10-06 f8e3feb
addresses #24
