### 2020-10-06 48c1cfb
fixes #16 and fixes #22
