### 2020-10-06 e9d4992
update viewing figures for Trin Tragula
