+++ man/schwarzschild-package.Rd
@@ -13,8 +13,8 @@ The DESCRIPTION file:
 
 Most of the package functionality is the definition of functions such as
 \code{schwarzschild()} which produce spacetime diagrams.  The functions
-are intended to be called by\code{inst/maker.R} which produces annotated
-and correctly-sized PDF files.
+are intended to be called by \code{inst/maker.R} which produces
+annotated and correctly-sized PDF files.
 
 \packageIndices{schwarzschild}
 }
