### 2020-10-07 6993171
make clean should remove png files
