+++ inst/Makefile
@@ -2,6 +2,6 @@ all :    maker.R
 	R CMD BATCH maker.R
 
 clean :
-	rm -f *.pdf *.svg *.jpg .RData maker.Rout
+	rm -f *.pdf *.svg *.jpg *.png .RData maker.Rout
 
 
