+++ DESCRIPTION
@@ -11,7 +11,7 @@ Description:
   mathematically accurate plotting routines; also creates animations
   of gravitational waves.  For special relativity, see the 'lorentz'
   package.
-Depends: gsl,deSolve,animation,grImport
+Depends: R (>= 3.0),gsl,deSolve,animation,grImport
 License: GPL (>= 2)
 Suggests: knitr, rmarkdown, testthat
 LazyLoad: yes
