### 2020-10-08 5d0c06f
more terse documentation
