+++ R/penrose_logistic.R
@@ -4,7 +4,7 @@ penrose_logistic <- function(colours=standard_colours, ...){
 
   penrose <- penrose_transform("logistic")
 
-  ## NB: norm looks good and tan sucks; cf penrose_black_hole
+  ## NB: norm looks good and this does not; cf penrose_black_hole
 
   ## set up axes:
   jj <- c(-1,1)
