### 2020-10-09 #91 @RobinHankin
Hello @IanHawke   @atrisovic  @whedon

thank you all for your comments, I have taken your suggestions on board and addressed them under https://github.com/RobinHankin/schwarzschild/issues/14 which is an umbrella issue that points to each of your observations as a separate repository issue.   The only comment I found confusing was "some of the sentences might be too informal or exaggerated for an academic paper", which left me feeling somewhat bewildered.  I have made changes here and there to the wording  but worry that I have misunderstood something.

@IanHawke  could please advise on my next action, and thanks again for your advice.  Best wishes, Robin
