### 2020-10-09 #91 @atrisovic
> thank you all for your comments, I have taken your suggestions on board and addressed them under [RobinHankin/schwarzschild#14](https://github.com/RobinHankin/schwarzschild/issues/14) which is an umbrella issue that points to each of your observations as a separate repository issue. The only comment I found confusing was "some of the sentences might be too informal or exaggerated for an academic paper", which left me feeling somewhat bewildered. I have made changes here and there to the wording but worry that I have misunderstood something.

Ok, nice! I will review these changes to check off the remaining points in the checklist above.

Sorry the comment was confusing - it was related to the sentence where you use personal pronouns which are not common in formal writing (ie, "as resources for **my own** astrophysics lectures at AUT, and secondly to support **my** Trin Tragula YouTube channel"). Also "heavily documented" seemed as an overstatement at the time. Would you like to change this before I review the text again?
