### 2020-10-12 #91 @RobinHankin
Hello @atrisovic  thanks for this clarification, I have now changed the wording as suggested in  commit

https://github.com/RobinHankin/schwarzschild/commit/51386ba9901648af16d2c6e207e1444d6222938e

Best wishes,  Robin
