### 2020-10-14 #91 @afolkest
Hi @RobinHankin

I tested the functionality and it looks good to me. It would perhaps be good to add one sentence in the introduction of the paper stating what the paper describes -- namely a visualization package for all the coordinate systems you do mention there. A brief comment on what method you use to solve the geodesic equation would be informative as well.

Also, I was wondering where I could see a list of dependencies, if any? Also, there is no release version.

Best,
Aasmund
