### 2020-10-14 RobinHankin/schwarzschild#14 @RobinHankin
User @afolkest makes a few suggestions under openjournals/jose-reviews#91 at

https://github.com/openjournals/jose-reviews/issues/91#issuecomment-708092146

I will open issues in here as I did above, linking here to issue #14,  and deal with them one by one.
