### 2020-10-18 #91 @RobinHankin
OK @afolkest, @atrisovic, @IanHawke @whedon, I have addressed all your comments at https://github.com/RobinHankin/schwarzschild/issues/14 and created
  tag 1.0-3  at https://github.com/RobinHankin/schwarzschild/releases/tag/1.0-3.  I hope this is satisfactory!  If not, let me know.  Best wishes, Robin
