### 2020-10-21 #91 @atrisovic
Hi again @RobinHankin,

I've now reviewed the package again and read the new version of the paper. The editor will probably better comment on these, but here are a few stylistic suggestions for the paper:

- The second sentence in the Statement of Need is a bit unclear. In particular, the part "not an object in its own right" is unclear.
- "GPL" -> General Public License
- Adding commas in a few places would help the reader:
    - "In computer terminology, one would need both the source code and the final image, but..."
    - "...which, as of October 2020, has..."
    - "...coordinate systems are available, including..."
- Why is plotting referred to as "side-effect", when that's the main functionality of the package?
    - "The package defines over twenty functions that are called for their side-effect of plotting a diagram of spacetime in the vicinity of a black hole." -> "The package defines over twenty functions that plot diagrams of spacetime in the vicinity of a black hole."
    - "Physical processes such as null geodesics and freely falling objects are simulated near a Schwarzschild black hole by the R functionality of the package; one side-effect is the the creation of usable and accurate PDF images." -> "Physical processes such as null geodesics and freely falling objects near a Schwarzschild black hole are simulated in R and plotted as usable and accurate PDF images."
- The Downloads section has only one sentence and an URL beyond the page width. I suggest moving the sentence in the previous section and hiding the URL.
    - For example: "Two example images ~~from the 19 in the package vignette~~ are shown below in low-resolution jpg format. ~~For vectorized PDF, see the package vignette.~~ All nineteen vectorized PDF images can be seen in the package vignette or online through [Microsoft Sharepoint](https://autuni-my.sharepoint.com/:f:/g/personal/rhankin_aut_ac_nz/EgX_IANsoOJDmTiH2i9_P20B6ksn9CMHf_TM31w5K3aITg?e=dPxAcp)."
    - Then remove the sentences: "For vectorized PDF version, see the package vignette or click the link below." **x2**
- jpg -> JPG format

Other than that, everything looks good to me, and all the boxes from the reviewers' list are now checked.
