### 2020-10-21 RobinHankin/schwarzschild#31 @RobinHankin
JOSE reviews "statement of need is unclear"
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

The second sentence in the Statement of Need is a bit unclear. In particular, the part "not an object in its own right" is unclear.

(issue #14 refers)
