### 2020-10-21 RobinHankin/schwarzschild#32 @RobinHankin
JOSE reviews GPL
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

"GPL" -> General Public License

(issue #14 refers)
