### 2020-10-21 RobinHankin/schwarzschild#33 @RobinHankin
JOSE reviews commas 
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

Adding commas in a few places would help the reader:
"In computer terminology, one would need both the source code and the final image, but..."
"...which, as of October 2020, has..."
"...coordinate systems are available, including..."

(issue #14 refers)
