### 2020-10-21 RobinHankin/schwarzschild#34 @RobinHankin
JOSE reviews "side effect" query
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

Why is plotting referred to as "side-effect", when that's the main functionality of the package?

(issue #14 refers)
