### 2020-10-21 RobinHankin/schwarzschild#35 @RobinHankin
JOSE review "over twenty functions"
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

"The package defines over twenty functions that are called for their side-effect of plotting a diagram of spacetime in the vicinity of a black hole." -> "The package defines over twenty functions that plot diagrams of spacetime in the vicinity of a black hole."

(issue #14 refers)
