### 2020-10-21 RobinHankin/schwarzschild#36 @RobinHankin
JOSE reviews "physical processes"
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

"Physical processes such as null geodesics and freely falling objects are simulated near a Schwarzschild black hole by the R functionality of the package; one side-effect is the the creation of usable and accurate PDF images." -> "Physical processes such as null geodesics and freely falling objects near a Schwarzschild black hole are simulated in R and plotted as usable and accurate PDF images."

(issue #14 refers)
