### 2020-10-21 RobinHankin/schwarzschild#37 @RobinHankin
JOSE reviews "only one sentence"
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

The Downloads section has only one sentence and an URL beyond the page width. I suggest moving the sentence in the previous section and hiding the URL.
For example: "Two example images from the 19 in the package vignette are shown below in low-resolution jpg format. For vectorized PDF, see the package vignette. All nineteen vectorized PDF images can be seen in the package vignette or online through Microsoft Sharepoint."
Then remove the sentences: "For vectorized PDF version, see the package vignette or click the link below." x2

(issue #14 refers)
