### 2020-10-21 RobinHankin/schwarzschild#38 @RobinHankin
JOSE reviews jpg -> JPG
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

jpg -> JPG format

(issue #14 refers)
