### 2020-10-21 RobinHankin/schwarzschild#34 @RobinHankin
The term "side-effect" is standard programming vernacular.  All R functions return a value but some functions do other things such as saving files or plotting a diagram.   Such operations are called _side-effects._  Functions such as `base::plot()` are usually executed primarily because of their side-effects [it actually returns `NULL`].  In the package, functions such as `schwarzschild()` and `kruskal()` have the same intended workflow.
