### 2020-10-21 fb19ed6
new version number for closing issue #14
