### 2020-10-21 #91 @RobinHankin
Hello  @atrisovic, thank you for your suggestions.  I have addressed all of them
under https://github.com/RobinHankin/schwarzschild/issues/14 and I think the paper reads much better as a result.  Tag 1.0-4  at https://github.com/RobinHankin/schwarzschild/releases/tag/1.0-4 refers.   What is the next step?  Best wishes, Robin
