### 2020-10-22 #91 @atrisovic
Ok, fantastic, you incorporated everything. Well, I'd still call the section "Examples" instead of "Downloads". And the last sentence in both of the figure captions could be now removed. (sorry for nitpicking!) :smile:

But yeah, we now wait for further instructions from @IanHawke.
