### 2020-10-22 RobinHankin/schwarzschild#14 @RobinHankin
Some comments from https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

I will open issues in here as I did above, linking here to issue #14, and deal with them one by one.
