### 2020-10-22 RobinHankin/schwarzschild#39 @RobinHankin
JOSE reviews: change  "Downloads" to "Examples"
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

 I'd still call the section "Examples" instead of "Downloads".

(issue #14 refers)
