### 2020-10-22 RobinHankin/schwarzschild#40 @RobinHankin
JOSE reviews: remove last sentence from figure captions
taken from @atrisovic, at  https://github.com/openjournals/jose-reviews/issues/91#issuecomment-713244181

And the last sentence in both of the figure captions could be now removed.

(issue #14 refers)
