### 2020-10-22 622f13c
fixes #40
