### 2020-10-22 63eed21
preparation for tag 1.0-5, issue #14 closed
