### 2020-10-22 #91 @RobinHankin
OK @atrisovic, thanks for these, no worries.  Happy to adjust the paper as suggested, see https://github.com/RobinHankin/schwarzschild/issues/14 and tagged at https://github.com/RobinHankin/schwarzschild/releases/tag/1.0-5

@IanHawke  I have changed the paper (and the vignette) in response to the comments by the referees and I await your instructions.  Best wishes,  Robin
