### 2021-05-23 #91 @labarba
Hi @RobinHankin — Apologies for the ball dropped here, and thank you for your enormous patience. I see from the review thread that thorough reviews were contributed, and you made changes to accommodate the reviewer requests. All checklist items are marked out. Let's get this past the finish line!

Could you now up the version, issue a tagged release in your repo, and make a Zenodo deposit? Report the version number here, as well as the Zenodo DOI, so we can run the commands to update that metadata for the submission. Then we can proceed to publication.
