### 2021-05-23 #91 @labarba
I see that there are no further commits on the repo after version 1.0-5, so I've set this as the version number.

We just need the Zenodo archive DOI now.
