### 2021-05-24 #91 @RobinHankin
hello Lorena, thank you for this!  Looking at the references, they are all
quite old and do not have DOIs.  I have uploaded to zenodo as you suggested
and it gave me 10.5281/zenodo.4782758

as the (provisional?) DOI.

Many thanks

Robin

***@***.***>

On Mon, May 24, 2021 at 12:38 AM Lorena A. Barba ***@***.***>
wrote:

> @RobinHankin <https://github.com/RobinHankin> — on checking your paper, I
> see that we will need you to add the DOIs for references that have them.
>
