### 2021-05-24 #91 @labarba
Hi @RobinHankin — I checked the Zenodo archive at https://doi.org/10.5281/zenodo.4782757 and you deposited the PDF of the JOSE paper. What we request is an archive of the submission _repository_ at https://github.com/RobinHankin/schwarzschild

This guide may be useful: https://guides.github.com/activities/citable-code/
