### 2021-05-24 #91 @RobinHankin
thank you for this Lorena, sorry for the confusion, I have now pointed to
release 1.0-5 on zenodo as you requested.

With best wishes

Robin

***@***.***>

On Mon, May 24, 2021 at 8:29 PM Lorena A. Barba ***@***.***>
wrote:

> Hi @RobinHankin <https://github.com/RobinHankin> — I checked the Zenodo
> archive at https://doi.org/10.5281/zenodo.4782757 and you deposited the
> PDF of the JOSE paper. What we request is an archive of the submission
> *repository* at https://github.com/RobinHankin/schwarzschild
>
> This guide may be useful:
> https://guides.github.com/activities/citable-code/
>
