### 2021-05-25 #91 @labarba
I checked the Zenodo archive at https://doi.org/10.5281/zenodo.4782757 and still see only the PDF of the paper there. The complete file set in the repository needs to be archived for persistence in the scholarly record. If you made a new Zenodo deposit, please post the link here.
