### 2021-05-25 #91 @RobinHankin
OK Lorena, thank you for your help here, the schwarzschild repo is now
visible on the zenodo page:

https://zenodo.org/record/4784248#.YKy61nUzZNg

Best wishes

Robin

***@***.***>

On Tue, May 25, 2021 at 7:00 PM Lorena A. Barba ***@***.***>
wrote:

> I checked the Zenodo archive at https://doi.org/10.5281/zenodo.4782757
> and still see only the PDF of the paper there. The complete file set in the
> repository needs to be archived for persistence in the scholarly record. If
> you made a new Zenodo deposit, please post the link here.
>
