### 2021-05-25 #91 @labarba
@RobinHankin – Almost there! We'd like to ask that you edit the metadata of the Zenodo deposit so the title and author list match the JOSE paper. (Zenodo grabs authors automatically from committers to your repo: it needs to be manually fixed.)
