### 2021-05-25 #91 @RobinHankin
OK Lorena, done!

https://zenodo.org/record/4784248

***@***.***>

On Tue, May 25, 2021 at 9:35 PM Lorena A. Barba ***@***.***>
wrote:

> @RobinHankin <https://github.com/RobinHankin> – Almost there! We'd like
> to ask that you edit the metadata of the Zenodo deposit so the title and
> author list match the JOSE paper. (Zenodo grabs authors automatically from
> committers to your repo: it needs to be manually fixed.)
>
